refactor(internal): move Tag out of types
